examples/installer: Break out partition number helper
